+++ packages/flutter_tools/lib/src/localizations/localizations_utils.dart
@@ -3,8 +3,10 @@
 // found in the LICENSE file.
 
 import 'package:meta/meta.dart';
+import 'package:yaml/yaml.dart';
 
 import '../base/file_system.dart';
+import '../base/logger.dart';
 import 'language_subtag_registry.dart';
 
 typedef HeaderGenerator = String Function(String regenerateInstructions);
@@ -285,4 +287,171 @@ String generateString(String value) {
     .replaceAll(backslash, r'\\');
 
   return "'$value'";
+}
+
+/// Typed configuration from the localizations config file.
+class LocalizationOptions {
+  const LocalizationOptions({
+    this.arbDirectory,
+    this.templateArbFile,
+    this.outputLocalizationsFile,
+    this.untranslatedMessagesFile,
+    this.header,
+    this.outputClass,
+    this.outputDirectory,
+    this.preferredSupportedLocales,
+    this.headerFile,
+    this.deferredLoading,
+    this.useSyntheticPackage = true,
+    this.areResourceAttributesRequired = false,
+  }) : assert(useSyntheticPackage != null);
+
+  /// The `--arb-dir` argument.
+  ///
+  /// The directory where all input localization files should reside.
+  final Uri arbDirectory;
+
+  /// The `--template-arb-file` argument.
+  ///
+  /// This URI is relative to [arbDirectory].
+  final Uri templateArbFile;
+
+  /// The `--output-localization-file` argument.
+  ///
+  /// This URI is relative to [arbDirectory].
+  final Uri outputLocalizationsFile;
+
+  /// The `--untranslated-messages-file` argument.
+  ///
+  /// This URI is relative to [arbDirectory].
+  final Uri untranslatedMessagesFile;
+
+  /// The `--header` argument.
+  ///
+  /// The header to prepend to the generated Dart localizations.
+  final String header;
+
+  /// The `--output-class` argument.
+  final String outputClass;
+
+  /// The `--output-dir` argument.
+  ///
+  /// The directory where all output localization files should be generated.
+  final Uri outputDirectory;
+
+  /// The `--preferred-supported-locales` argument.
+  final List<String> preferredSupportedLocales;
+
+  /// The `--header-file` argument.
+  ///
+  /// A file containing the header to prepend to the generated
+  /// Dart localizations.
+  final Uri headerFile;
+
+  /// The `--use-deferred-loading` argument.
+  ///
+  /// Whether to generate the Dart localization file with locales imported
+  /// as deferred.
+  final bool deferredLoading;
+
+  /// The `--synthetic-package` argument.
+  ///
+  /// Whether to generate the Dart localization files in a synthetic package
+  /// or in a custom directory.
+  final bool useSyntheticPackage;
+
+  /// The `required-resource-attributes` argument.
+  ///
+  /// Whether to require all resource ids to contain a corresponding
+  /// resource attribute.
+  final bool areResourceAttributesRequired;
+}
+
+/// Parse the localizations configuration options from [file].
+///
+/// Throws [Exception] if any of the contents are invalid. Returns a
+/// [LocalizationOptions] with all fields as `null` if the config file exists
+/// but is empty.
+LocalizationOptions parseLocalizationsOptions({
+  @required File file,
+  @required Logger logger,
+}) {
+  final String contents = file.readAsStringSync();
+  if (contents.trim().isEmpty) {
+    return const LocalizationOptions();
+  }
+  final YamlNode yamlNode = loadYamlNode(file.readAsStringSync());
+  if (yamlNode is! YamlMap) {
+    logger.printError('Expected ${file.path} to contain a map, instead was $yamlNode');
+    throw Exception();
+  }
+  final YamlMap yamlMap = yamlNode as YamlMap;
+  return LocalizationOptions(
+    arbDirectory: _tryReadUri(yamlMap, 'arb-dir', logger),
+    templateArbFile: _tryReadUri(yamlMap, 'template-arb-file', logger),
+    outputLocalizationsFile: _tryReadUri(yamlMap, 'output-localization-file', logger),
+    untranslatedMessagesFile: _tryReadUri(yamlMap, 'untranslated-messages-file', logger),
+    header: _tryReadString(yamlMap, 'header', logger),
+    outputClass: _tryReadString(yamlMap, 'output-class', logger),
+    outputDirectory: _tryReadUri(yamlMap, 'output-dir', logger),
+    preferredSupportedLocales: _tryReadStringList(yamlMap, 'preferred-supported-locales', logger),
+    headerFile: _tryReadUri(yamlMap, 'header-file', logger),
+    deferredLoading: _tryReadBool(yamlMap, 'use-deferred-loading', logger),
+    useSyntheticPackage: _tryReadBool(yamlMap, 'synthetic-package', logger) ?? true,
+    areResourceAttributesRequired: _tryReadBool(yamlMap, 'required-resource-attributes', logger) ?? false,
+  );
+}
+
+// Try to read a `bool` value or null from `yamlMap`, otherwise throw.
+bool _tryReadBool(YamlMap yamlMap, String key, Logger logger) {
+  final Object value = yamlMap[key];
+  if (value == null) {
+    return null;
+  }
+  if (value is! bool) {
+    logger.printError('Expected "$key" to have a bool value, instead was "$value"');
+    throw Exception();
+  }
+  return value as bool;
+}
+
+// Try to read a `String` value or null from `yamlMap`, otherwise throw.
+String _tryReadString(YamlMap yamlMap, String key, Logger logger) {
+  final Object value = yamlMap[key];
+  if (value == null) {
+    return null;
+  }
+  if (value is! String) {
+    logger.printError('Expected "$key" to have a String value, instead was "$value"');
+    throw Exception();
+  }
+  return value as String;
+}
+
+List<String> _tryReadStringList(YamlMap yamlMap, String key, Logger logger) {
+  final Object value = yamlMap[key];
+  if (value == null) {
+    return null;
+  }
+  if (value is String) {
+    return <String>[value];
+  }
+  if (value is Iterable) {
+    return value.map((dynamic e) => e.toString()).toList();
+  }
+  logger.printError('"$value" must be String or List.');
+  throw Exception();
+}
+
+// Try to read a valid `Uri` or null from `yamlMap`, otherwise throw.
+Uri _tryReadUri(YamlMap yamlMap, String key, Logger logger) {
+  final String value = _tryReadString(yamlMap, key, logger);
+  if (value == null) {
+    return null;
+  }
+  final Uri uri = Uri.tryParse(value);
+  if (uri == null) {
+    logger.printError('"$value" must be a relative file URI');
+  }
+  return uri;
 }
